Refactor SQL query and add transaction time index
Rather than the nested sets of sub-SELECTs that were extremely difficult
to parse and often ended up being superflouus, use targeted explicit
JOINs to pull together the tables in a hopefully more logical manner -
or at least a manner that lends itself to being maintained by multiple
users.

Also move towards an SQL function rather than keeping the complex query
inside the Python code itself. We're not quite there yet, as we still
need to dynamically add the clause that optionally restricts the query
to a given institution, and that keeps alive the possibility for SQL
injection - but we have used placeholders properly for the rest of the
variables in the getRecommendations() function. Likely direction to take
this is to have two versions of the SQL function, with one accepting an
array of the institution names so that it can do the lookup inline (via
a CTE). Then we just call the appropriate function, depending on whether
we have an array or not, and we can use placeholders properly in the
Python code.

Signed-off-by: Dan Scott <dan@coffeecode.net>