Changed item table in the database so that all of the information know about items from the first data load (title, author, publisher, year, URL) is included.  Changed the loading procedure to match.
Changed transaction loading call so it uses old_load_transactions_fromfile because there was some kind of problem with the new function.

Various minor changes to the index.html page and the slider stuff, including bringing the CSS file into the code base so it's easy to change.

Redid the test data so there are more transactions.  I just copied one small set of transactions two times and upped the timestamps.