Use a safer (but slower) data loading method for items
In load_items_fromfile() (in splurge.py), load the items information
into the database line by line instead of with copy.  This solves
the problem of the entire load breaking when there's a duplicate
item/ISBN.  Instead of the whole thing failing, that one line is
thrown out and then it continues.

This is, unfortunately, much slower, but perhaps that can be fixed.

Signed-off-by: Dan Scott <dscott@laurentian.ca>