Simplify README for PostgreSQL setup and data load
On most distributions, using the "-h localhost" flag will enable you to
log in as the desired user without having to concern yourself with peer
authentication. This also matches how we're authenticating in the
application itself.

In addition, provide a clear structure for the location of the items and
transactions files.

Signed-off-by: Dan Scott <dscott@laurentian.ca>